+++ R/atmospheric_attenuation.R
@@ -31,7 +31,7 @@
 #'atmospheric_attenuation(f = 20000, temp = 20, RH = 90, p = 88000, formula = 1)
 #' }
 #' 
-#' @author Marcelo Araya-Salas (\email{marceloa27@@gmail.com}) 
+#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr}) 
 #' @references {
 #' Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2
 #' }