+++ R/master_sound_file.R
@@ -42,7 +42,7 @@
 #' # Rraven::exp_raven(master.sel.tab, path = tempdir(), file.name = "example_master_selection_table")
 #' }
 #' 
-#' @author Marcelo Araya-Salas (\email{marceloa27@@gmail.com})
+#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 #' @seealso \code{\link[Rraven]{exp_raven}}
 #' @references {
 #' Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2