+++ R/noise_profile.R
@@ -44,7 +44,7 @@
 #'  noise_profile(X = pe, mar = 0.01, pb = FALSE, noise.ref = "adjacent")
 #' }
 #' 
-#' @author Marcelo Araya-Salas (\email{marceloa27@@gmail.com})
+#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
 #' @seealso \code{\link{excess_attenuation}}
 #' @references {
 #' Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.2.